Create a postListItem component to avoid merge conflicts
